Move functionality out of AlbumPage into AlbumOverview
